Refactored test and added test for sign up template
