--- accounts/tests/test_admin.py
@@ -1,40 +1,0 @@
-from django.test import TestCase, Client
-from django.contrib.auth import get_user_model
-from django.urls import reverse
-
-
-class AdminSiteTests(TestCase):
-
-    def setUp(self):
-        self.client = Client()
-        self.admin_user = get_user_model().objects.create_superuser(
-            email='test_admin@gmail.com',
-            password='adminpassword'
-        )
-        self.client.force_login(self.admin_user)
-        self.user = get_user_model().objects.create_user(
-            email='test@gmail.com',
-            password='userpassword',
-            name='User for test'
-        )
-
-    def test_users_listed(self):
-        """test that users are listed on user page"""
-        url = reverse('admin:accounts_customuser_changelist')
-        res = self.client.get(url)
-        self.assertContains(res, self.user.name)
-        self.assertContains(res, self.user.email)
-
-    def test_user_change_page(self):
-        """Test that the user edit page works"""
-        url = reverse('admin:accounts_customuser_change', args=[self.user.id])
-        res = self.client.get(url)
-
-        self.assertEqual(res.status_code, 200)
-
-    def test_create_user_page(self):
-        """Test that the create user page works"""
-        url = reverse('admin:accounts_customuser_add')
-        res = self.client.get(url)
-
-        self.assertEqual(res.status_code, 200)
