+++ accounts/tests/test_user.py
@@ -1,5 +1,6 @@
-from django.test import TestCase
+from django.test import TestCase, Client
 from django.contrib.auth import get_user_model
+from django.urls import reverse
 
 
 class Custom_User_Model_Test(TestCase):
@@ -46,3 +47,52 @@ class Custom_User_Model_Test(TestCase):
         self.assertTrue(admin_user.is_active)
         self.assertTrue(admin_user.is_superuser)
         self.assertTrue(admin_user.is_staff)
+
+class AdminSiteTests(TestCase):
+
+    def setUp(self):
+        self.client = Client()
+        self.admin_user = get_user_model().objects.create_superuser(
+            email='test_admin@gmail.com',
+            password='adminpassword'
+        )
+        self.client.force_login(self.admin_user)
+        self.user = get_user_model().objects.create_user(
+            email='test@gmail.com',
+            password='userpassword',
+            name='User for test'
+        )
+
+    def test_users_listed(self):
+        """test that users are listed on user page"""
+        url = reverse('admin:accounts_customuser_changelist')
+        res = self.client.get(url)
+        self.assertContains(res, self.user.name)
+        self.assertContains(res, self.user.email)
+
+    def test_user_change_page(self):
+        """Test that the user edit page works"""
+        url = reverse('admin:accounts_customuser_change', args=[self.user.id])
+        res = self.client.get(url)
+
+        self.assertEqual(res.status_code, 200)
+
+    def test_create_user_page(self):
+        """Test that the create user page works"""
+        url = reverse('admin:accounts_customuser_add')
+        res = self.client.get(url)
+
+        self.assertEqual(res.status_code, 200)
+
+
+class SignUpPageTests(TestCase):
+
+    def setUp(self):
+        url = reverse('signup')
+        self.response = self.client.get(url)
+
+    def test_signup_template(self):
+        """Test that singup template is rendered correctly"""
+        self.assertTemplateUsed(self.response, 'registration/signup.html')
+        self.assertContains(self.response, 'Sign Up')
+        self.assertNotContains(self.response, 'I am not on this page!')
